DEV fix processing of some commands and exec events
