+++ device_descriptions.cpp
@@ -1706,3 +1706,32 @@ uint8_t DDF_GetSubDeviceOrder(const QString &type)
 
     return SUBDEVICE_DEFAULT_ORDER;
 }
+
+/*! Creates a unique Resource handle.
+ */
+Resource::Handle R_CreateResourceHandle(const Resource *r, size_t containerIndex)
+{
+    Q_ASSERT(r->prefix() != nullptr);
+    Q_ASSERT(!r->item(RAttrUniqueId)->toString().isEmpty());
+
+    Resource::Handle result;
+    result.hash = qHash(r->item(RAttrUniqueId)->toString());
+    result.index = containerIndex;
+    result.type = r->prefix()[1];
+    result.order = 0;
+
+    Q_ASSERT(result.type == 's' || result.type == 'l' || result.type == 'd' || result.type == 'g');
+    Q_ASSERT(isValid(result));
+
+    if (result.type == 's' || result.type == 'l')
+    {
+        const ResourceItem *type = r->item(RAttrType);
+        if (type)
+        {
+            result.order = DDF_GetSubDeviceOrder(type->toString());
+        }
+    }
+
+    return result;
+}
+
