+++ device_descriptions.h
@@ -268,5 +268,6 @@ private:
 
 void DDF_AnnoteZclParse1(int line, const char* file, const Resource *resource, ResourceItem *item, quint8 ep, quint16 clusterId, quint16 attributeId, const char *eval);
 const DeviceDescription::Item &DDF_GetItem(const ResourceItem *item);
+Resource::Handle R_CreateResourceHandle(const Resource *r, size_t containerIndex);
 
 #endif // DEVICEDESCRIPTIONS_H
