+++ resource.h
@@ -333,6 +333,7 @@ public:
     const char *suffix = RInvalidSuffix;
     qint64 validMin = 0;
     qint64 validMax = 0;
+    quint16 flags = 0;
 };
 
 class Resource;
@@ -344,6 +345,7 @@ extern const ResourceItemDescriptor rInvalidItemDescriptor;
 
 class ResourceItem
 {
+public:
     enum ItemFlags
     {
         FlagNeedPushSet     = 0x01, // set after a value has been set
@@ -354,7 +356,6 @@ class ResourceItem
         FlagImplicit        = 0x20  // the item is always present for a specific resource type
     };
 
-public:
     enum ValueSource
     {
         SourceUnknown,
@@ -588,7 +589,6 @@ bool R_SetValueEventOnSet(Resource *r, const char *suffix, const V &val, Resourc
 
 uint8_t DDF_GetSubDeviceOrder(const QString &type);
 QLatin1String R_DataTypeToString(ApiDataType type);
-Resource::Handle R_CreateResourceHandle(const Resource *r, size_t containerIndex);
 inline bool isValid(Resource::Handle hnd) { return hnd.hash != 0 && hnd.index < UINT16_MAX && hnd.type != 0; }
 inline bool operator==(Resource::Handle a, Resource::Handle b) { return a.hash == b.hash && a.type == b.type; }
